首页tab调整，详情地图处理，定向乐跑提示处理

+++ app/src/main/java/com/lptiyu/tanke/activities/directionrun/DirectionRunContact.java
@@ -27,8 +27,8 @@ public class DirectionRunContact {
 
         void getRunLine(long gameId);
 
-        void runSignUp(long gameId, long pointId, long recordId, long distance);
+        void runSignUp(long gameId, long pointId, long recordId, double distance);
 
-        void stopRun(long gameId, long recordId, long distance);
+        void stopRun(long gameId, long recordId, double distance);
     }
 }
